backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
google-api-python-client
google-generativeai
gunicorn
